code/05-eval-pipeline/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
